Move testInstant into SQLTestBase
