+++ src/test/java/io/vertx/ext/asyncsql/PostgreSQLTest.java
@@ -21,7 +21,6 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.ext.sql.ResultSet;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
-import java.time.Instant;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -129,29 +128,4 @@ public class PostgreSQLTest extends AbstractTestBase {
     });
   }
 
-  @Test
-  public void testInstant(TestContext context) {
-    Async async = context.async();
-    client.getConnection(ar -> {
-      ensureSuccess(context, ar);
-      conn = ar.result();
-      conn.execute("DROP TABLE IF EXISTS test_table", ar1 -> {
-        ensureSuccess(context, ar1);
-        conn.execute("CREATE TABLE test_table (instant TIMESTAMP)", ar2 -> {
-          ensureSuccess(context, ar2);
-          Instant now = Instant.now();
-          conn.queryWithParams("INSERT INTO test_table (instant) VALUES (?)", new JsonArray().add(now), ar3 -> {
-            ensureSuccess(context, ar3);
-            conn.query("SELECT instant FROM test_table", ar4 -> {
-              ensureSuccess(context, ar4);
-              // timestamps with out time zone are returned as strings, so we must compare to the original instant
-              // ignoring the timezone offset (meaning ignore everything after char 23)
-              context.assertEquals(ar4.result().getResults().get(0).getString(0), now.toString().substring(0, 23));
-              async.complete();
-            });
-          });
-        });
-      });
-    });
-  }
 }
