+++ src/test/java/io/vertx/ext/asyncsql/SQLTestBase.java
@@ -29,6 +29,7 @@ import io.vertx.ext.unit.TestContext;
 import org.junit.Test;
 
 import java.math.BigDecimal;
+import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
 
@@ -600,6 +601,40 @@ public abstract class SQLTestBase extends AbstractTestBase {
         }
       });
     });
+  }
+
+  @Test
+  public void testInstant(TestContext context) {
+    Async async = context.async();
+    client.getConnection(ar -> {
+      ensureSuccess(context, ar);
+      conn = ar.result();
+      conn.execute("DROP TABLE IF EXISTS test_table", ar1 -> {
+        ensureSuccess(context, ar1);
+        conn.execute("CREATE TABLE test_table (instant TIMESTAMP)", ar2 -> {
+          ensureSuccess(context, ar2);
+          Instant now = Instant.now();
+          conn.queryWithParams("INSERT INTO test_table (instant) VALUES (?)", new JsonArray().add(now), ar3 -> {
+            ensureSuccess(context, ar3);
+            conn.query("SELECT instant FROM test_table", ar4 -> {
+              ensureSuccess(context, ar4);
+              // timestamps with out time zone are returned as strings, so we must compare to the original instant
+              // ignoring the timezone offset (meaning ignore everything after char 23)
+              compareInstantStrings(
+                  context,
+                  ar4.result().getResults().get(0).getString(0),
+                  now.toString().substring(0, 23)
+              );
+              async.complete();
+            });
+          });
+        });
+      });
+    });
+  }
+
+  protected void compareInstantStrings(TestContext context, String result, String expected) {
+    context.assertEquals(result, expected);
   }
 
   protected void setSqlModeIfPossible(Handler<Void> handler) {
